perf(config): cache OXYROUTE_* env flags on first use
Read OXYROUTE_MAX_BODY_BYTES and OXYROUTE_DEBUG once via OnceLock instead
of per-request env lookups; isolate env-mutating tests in subprocesses.
