+++ tests/test_handler_errors_500.py
@@ -46,22 +46,43 @@ def test_500_no_exception_text_by_default() -> None:
 
 
 def test_500_includes_detail_when_debug_env() -> None:
-    os.environ["OXYROUTE_DEBUG"] = "1"
-    app = _make_app()
-    try:
+    """Isolated process: ``OXYROUTE_DEBUG`` is cached on first error path."""
+    import subprocess
+    import sys
+    from pathlib import Path
 
-        async def _run() -> None:
-            transport = httpx.ASGITransport(app=asgi_test_app(app))
-            async with httpx.AsyncClient(transport=transport, base_url="http://t") as c:
-                r = await c.get("/boom")
-            assert r.status_code == 500
-            data = json.loads(r.text)
-            assert "detail" in data
-            assert "super_secret_token" in data["detail"] or "ValueError" in data["detail"]
+    repo = Path(__file__).resolve().parents[1]
+    code = """
+import asyncio
+import json
+import os
 
-        asyncio.run(_run())
-    finally:
-        os.environ.pop("OXYROUTE_DEBUG", None)
+import httpx
+from oxyroute import App
+from tests._rsgi_test_transport import asgi_test_app
+
+os.environ["OXYROUTE_DEBUG"] = "1"
+
+app = App()
+
+@app.get("/boom")
+def boom() -> str:
+    raise ValueError("super_secret_token")
+
+async def main() -> None:
+    transport = httpx.ASGITransport(app=asgi_test_app(app))
+    async with httpx.AsyncClient(transport=transport, base_url="http://t") as c:
+        r = await c.get("/boom")
+    assert r.status_code == 500
+    data = json.loads(r.text)
+    assert "detail" in data
+    assert "super_secret_token" in data["detail"] or "ValueError" in data["detail"]
+
+asyncio.run(main())
+"""
+    env = os.environ.copy()
+    env["PYTHONPATH"] = str(repo)
+    subprocess.run([sys.executable, "-c", code], env=env, check=True, cwd=repo)
 
 
 def test_dependency_error_returns_500() -> None:
